Add files via upload
My solution to an assignment in my Data Structures and Algorithms course. More information about this assignment are in the a4.pdf file. My implementation of the solution is in the ParisMetro.java file. A more efficient is definitely possible mainly by better choice of the data structures. This was a quick and dirty solution to that assignment. Constructive comments always welcome.